CI(GitHub): add tests to validate GitHub response
A new set of tests replaces the octokit endpoint with a wrapper to
allow any response to be returned, allowing GitHubGlue checks to
be fully and easily tested.

There are two sets of tests.  The first is using the octokit schemas
that requires extensive boilerplate data.  The second is schemas
limited to the data GitHubGlue is interested in.

Signed-off-by: Chris. Webster <chris@webstech.net>