Add a class designed to process the "What's cooking" emails
One aspect of Git's contribution process that often puzzles irregular
contributors is that there seem not to be any updates on their patches,
even if the branch was picked up. The reason for that puzzlement is that
those updates are not sent in-thread, but as part of the "What's
cooking" emails that are sent out roughly twice a week.

With this new code, we can start to address this puzzlement by keeping
contributors in the loop, in particular when the Git maintainer expects
a "reroll" (read: a new iteration of the patch series).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>