Add enhanced linting for new contributors
For first time contributors, extra checks will be done to
validate the PR.  A label will be added to flag the PR for further
review in many instances.  The label can then be removed by any
reviewer.  The intent is to NOT close PRs that may be valid.  Some
PRs will be closed based on the content.

Signed-off-by: Chris. Webster <chris@webstech.net>