Let contributors know about their What's Cooking status
It is frequently confusing contributors who are new to the project when
they do not hear anything about the status of their contribution for
weeks.

Sometimes, the reason is that they are expected to change something, but
the reminder to do so is only sent out via the What's Cooking mails
(which contributors won't receive unless they subscribe to the fire hose
that is the Git mailing list).

Let's be helpful and mirror that status update back to the PR.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>