Rename `pr-test` and activate it for PRs and all pushes
This maintainer relied on the tests to give us an idea whether anything
breaks badly if a PR is merged.

However, this workflow did not even run as part of
https://github.com/gitgitgadget/gitgitgadget/pull/497 and sure enough,
it introduced a rather big regression: the Windows job is now thoroughly
broken.

Let's instead run it on every PR and on every push (except
Dependabot's, as they do not need to trigger _two_ runs, one for a push
and one for a Pull Request).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>