test: use internal SMTP server
Use package `test-smtp-server` to run an internal SMTP server.  This
will remove the dependency on a third party and may be a little
faster.

This is an attempt to resolve the `ConnectionTimeoutError: connection
timed out. timeout = 3000 ms` errors.  These errors occur randomlyn
and break tests.  Recently, a popular fake SMTP server had a full
day outage.

Signed-off-by: Chris. Webster <chris@webstech.net>