Do not even update the local notes ref upon `/preview`
A large part of the `generateAndSend()` method is concerned with
updating the metadata that is stored in the `refs/notes/gitgitgadget`
ref, which holds GitGitGadget's state.

However, when sending a preview, we do not actually want to update this
metadata: no new iteration has _actually_ been sent yet.

Previously, this worked fine because while we updated the notes ref
locally, we actually never sent it to the remote repository. And the
next time we handled a `/submit` or a `/preview`, we force-fetched the
notes ref, which would overwrite the previous changes.

However, as reported by Lessley Dennington, this strategy failed in
https://github.com/gitgitgadget/git/pull/1050#issuecomment-985770245: a
preview was sent, but the notes ref was pushed _somehow_ nevertheless.
The best guess is that _another_ automated run updated the notes ref
independently, but did not force-fetch it first (GitGitGadget's
automated runs happen on a dedicated, persistent build agent, to avoid
having to fetch megabytes upon megabytes just to clone git.git and
lore.git, therefore one automated run can influence the next automated
run by mistake).

Let's safeguard the code by not even updating the local notes ref when
processing a `/preview` command.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>