update-open-prs: be graceful about a missing Message-ID for a patch
It should not happen, of course, that GitGitGadget sent off a patch
series and then fails to remember the Message-ID of the sent mail.

However, this just happened.

Instead of barfing on this, try our best to reconstruct the Message-ID
for the patch and store it in the internal database.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>