Add some white-space to long lines in the commit-lint tests
We are about to introduce an exception to the long lines linter: any
long line that simply cannot be wrapped at white-space (because there is
no white-space to wrap at) is exempt from the check.

Let's adjust the existing test cases that want to verify the long lines
linter: introduce white-space in the appropriate places.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>