Exempt un-wrappable long lines from the long line linter
We have a linter that verifies that commit messages do not contain long
lines. However, ny long line that simply cannot be wrapped at
white-space (because there is no white-space to wrap at) should be
exempt from the check.

Let's make it so.

This is intended to allow commit messages to contain lines like these:

[2] https://lore.kernel.org/git/CABPp-BH9tju7WVm=QZDOvaMDdZbpNXrVWQdN-jmfN8wC6YVhmw@mail.gmail.com/

Requested by Elijah Newren in
https://github.com/gitgitgadget/git/pull/1113#issuecomment-1008229286

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>