update-commit-mappings: do force-fetch upstream branches again
In 2c0a435 (extensible: introduce product-config feature, 2022-02-06),
we refactored the code to make the names of upstream branches
configurable. By mistake, we omitted the `+` that forces updates to the
corresponding remote-tracking branches in case of non-fast-forwards. But
we do have to expect that e.g. `seen` (and occasionally, `next`) need to
be force-fetched.

Noticed in
https://dev.azure.com/gitgitgadget/git/_build/results?buildId=128583

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>