Fix parsing of the `References:` header
When we replaced our home-grown mbox parser with a version that uses the
`mailparser`, we did not adjust the regex when parsing the `References:`
header.

As a consequence, our parser got confused when the references stretched
over multiple lines (because the regex did not expect Carriage Returns
nor Line Feeds).

Let's fix this.

This fixes https://github.com/gitgitgadget/gitgitgadget/issues/904.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>