ci-helper: do error out upon problems
We recently had a problem where Google stopped allowing "less-secure
access" by third-party apps, forcing us to switch to using app
passwords.

The biggest problem, however, was that the Pipeline runs were marked as
_successful_, even if the run should have been marked as failure because
the patch series was not sent successfully.

The reason is that caught the exception (and displayed an error) without
actually erroring out. This patch fixes that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>