MailArchiveHelper: avoid invalid links to the mail archive
The "What's cooking" base URL might end in a slash, or in a suffix like
`?messageId=` or something like that, but it is important that the
message Id can be appended as-is.

The way we configured it, it should have ended in a slash, but did not,
and we constructed invalid links to the "What's cooking" mails.

Let's not do that.

Note: For other mails, we assumed that the base URL _should_ end in a
slash but does not, and explicitly inserted a slash.

Let's not do that, either.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>