Format checkboxes
GitHub task lists are represented in HTML as unordered lists with
checkboxes.  These checkboxes are currently ignored when converting the
HTML to text.  This affects PRs with task lists (e.g.
https://github.com/git/git/pull/1359); the resulting text list has no
indication anymore whether an item was checked (done) or not.

Format checkboxes by turning checked ones into "[x]" and unchecked
ones into "[ ]" to retain that information during the conversion.

Helped-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: René Scharfe <l.s.r@web.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>