CIHelper: make the `config` a mandatory, explicit constructor parameter
It was a bit hard for me to follow the code, getting confused by two
separate `getConfig()` functions that do different things (the one from
`gitgitgadget-config` initializes the default config, the one from
`project-config` gets a singleton that might still be uninitialized).

Let's unconfuse things by making the config a non-optional parameter of
the constructor.

Note: This commit is only concerned with the CIHelper class, because we
will use it from light-weight GitHub Actions in the future. The rest of
the code really needs to use the singleton up until the time when we'll
get around to The Big Refactor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>