Install `ncc`
Preparing for turning GitGitGadget into a bunch of GitHub Actions, we
install `ncc` which will allow us to bundle a minimal set of Javascript
files and publish that via a tag.

The idea is to run `ncc build -s -m lib/ci-helper.ts -o dist/` to obtain
a minimized `dist/index.js` that has no additional dependencies, and
then have subdirectories (e.g. `update-prs/`) with the individual
`action.yml` and `index.js` that starts with the following line

	const { CIHelper } = require('../dist/index.js')

This will allow us to publish a single tag that backs the various
functionalities as different GitHub Actions.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>