misc-helper: refactor the `set-app-token` code
The idea is to move the `set()` function to CIHelper, to make it
reusable. Before that, however, we must make sure that it is generic
enough, e.g. by allowing to pass the App's private key as a parameter
(instead of expecting it in the environment or in the Git config).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>