Add the `gitgitgadget/init` GitHub Action
The idea is to let `ncc` turn the CIHelper code into a dependency-free
`dist/index.js` and then remove all files other than `*/action.yml` and
`*/index.js`, commit and tag that. That's the GitHub Action, ready to
be used in GitHub workflows.

Since `ncc` produces an ECMAScript module, we cannot use the old-style
`require()` call but need to use `await import()` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>