Remove support for the CLI mode
The GitGitGadget project originates from a Unix Shell script
named `mail-patch-series.sh` which is still available at
https://github.com/dscho/mail-patch-series. This script was first
converted to Typescript, and subsequently into a GitHub App whose
compute-intensive part is performed by Azure Pipelines.

The Typescript variant of the original command-line script is still
in the code-base, as is the support code only needed by that script
(such as extracting patch series' metadata from a tag). However,
this script is no longer used (not even by me, I turned back to the
Unix Shell variant for use cases not yet supported by GitGitGadget,
such as contributing to Cygwin).

Therefore, let's get rid of the script as well as the support code
needed only by that script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>