WELCOME: add a note about single-commit PRs
For single-commit PRs, GitHub prefills the PR description with the
content of the commit message.

First time Git contributors using Gitgitgadget for a single-commit PR
often keep this automatically generated PR description as-is, which thus
results in the patch being sent to the mailing list with identical text
in the commit message and below the three-dash line, which increases the
signal-to-noise ratio for reviewers.

Adjust the welcoming message in res/WELCOME.md to distinguish
single-commit from multiple-commits PRs. Do that in the sentence near
the top of the message, in the hope that it is read by at least some
contributors.

This partly adresses:
https://github.com/gitgitgadget/gitgitgadget/issues/340