Prepare for upgrading Git
As of v2.41.0-rc0~89^2 (e-mail workflow: Message-ID is spelled with ID
in both capital letters), the output of `format-patch` has changed.

Since the original format works well for GitGitGadget, let's force that.

An alternative would be to accept the new format, but it is unclear what
consequences that would have, whereas staying with the original format
is safe.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>