WELCOME.md: warn new users against copy/pastign CC lists
Copy/pasting Markdown-formatted CC lists from one GGG PR to another can
result in a broken CC list which needs manual fixing from the
maintainer. So warn users about this issue early on for all first-time
contributors.

See [1] for a discussion of some of the ways the community would like to
mitigate this issue (one of which is the motivation for this change).

[1] https://lore.kernel.org/git/xmqqle2o26jy.fsf@gitster.g/