eslint: convert to new flat config format
Convert from `.eslintrc` to `eslint.config` file format.  Remove the old config and ignore files.  The ignore was not needed anyway.

Replaced the trojan detection with `eslint-plugin-security`, which includes other checks.  Added `typescript-eslint` for ts checks.

Added `@stylistic` for lints moved out of eslint.  Some of these were intended to override prettier but that is not working.

Set `prettier` to warn.  It is too opinionated to be allowed to error.

Signed-off-by: Chris. Webster <chris@webstech.net>