misc-helper: make use of the modern `commander` paradigm
Nowadays, sub-commands are meant to be implemented via `.action()`
callbacks instead of `if ... else if ... else` chains.

The only major complication is that some of `misc-helper`'s sub-commands
take optional arguments _at the beginning_: `get-pr-meta`, for example,
takes an optional repository owner as first argument, followed by a
required PR number or URL. This is a scenario for which `commander` is
not prepared; We need to sub-class its `Command` class to work around
that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>