PatchSeries: do not pretend to publish a branch
In the olden days, when GitGitGadget was still `mail-patch-series.sh`,
the source of truth was the local repository where it was run, and as
such, the branch needed to be pushed to the remote to make it, well,
public.

However, in the current setup, the source of truth is the
gitgitgadget/git (or git/git) repository's Pull Request, i.e. the branch
is already published.

Therefore, this branch publishing logic should have been removed in
6d397b0f (Remove support for the CLI mode, 2023-12-31)... better late
than never...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>