WIP Action: remove all but the absolutely required files
This trick was performed by:

	git rm -r -- \* ':(exclude)dist/' ':(exclude)init/'

Note that we also need to move `res/` into `dist/`, and to initialize a
bare-bones `package.json` so that the GitHub Action runs as an ESModule.

My current idea is to do this whenever anything was pushed to `main`, in
the `v1` branch, in a fake merge. The idea being that `v1` always
contains just the Action's files, updated via merge commits whose 2nd
parent commit is the `main` branch. Something like this should
illustrate the strategy well:

	if git rev-parse --verify refs/remotes/origin/v1
	then
		# Good, `v1` exists already
		git switch -t origin/v1
	else
		# Need to create `v1` from scratch
		git switch --orphan v1
	fi &&

	# Check out the current files of `main` (without moving `HEAD`)
	git checkout origin/main -- . &&

	# Build dist/
	npm ci &&
	npm run build-dist &&

	# Now, add the generated files
	git add dist/{index.js{,.map},sourcemap-register.*} &&

	# Now, move `res/`
	git mv res dist/ &&

	# Remove everything that is not needed for the Action
	git rm -r -- \* ':(exclude)dist/' ':(exclude)init/' &&

	# Add bare-bones `package.json`
	printf '{\n    "type": "module"\n}\n' >package.json &&
	git add package.json &&

	# Now make that fake merge commit
	tree=$(git write-tree) &&
	msg="Sync with main ($(git show -s --pretty=reference origin/main))" &&
	commit=$(git commit-tree -m "$msg" $tree -p HEAD -p origin/main) &&
	git reset $commit &&

	# Verify that this committed all the files
	test -z "$(git ls-files --modified --other)"

Obviously this trick should be performed in a GitHub workflow that
triggers whenever there is a push to `main`.

This workflow would then also be responsible for adding an appropriate
tag (v1.<n>, where <n> increments the latest tag by one).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>