pushNotesRef: allow for concurrent operations
GitGitGadget's state lives in a JSON object at the tip of the
`refs/notes/gitgitgadget` ref at https://github.com/gitgitgadget/git.

When GitGitGadget's automation runs, it typically modifies that, and
then pushes the new state.

In case of concurrent runs, this would cause a problem (which is why we
currently prevent that). But not for long! This here commit uses the
`notesSync()` function that was introduced _just_ for the purpose of
allowing concurrent runs, by merging whatever changed upstream and then
pushing again.

To allow for multiple concurrent contenders, let's try that a couple of
times before giving up, waiting for an increasingly long time.

This functionality is now made available via the shiny new
`GitNotes.push(url)` method.

This method will henceforth be used by the `GitGitGadget.pushNotesRef()`
method to guarantee robustness for concurrent GitGitGadget workflows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>