TODOs
I'd like the following `misc-helper` commands to execute
`GitGitGadget.pushNotesRef()` in a `finally` clause, to benefit from the
just-introduced support for concurrent GitGitGadget workflow runs:

- update-open-prs

- update-commit-mappings

- handle-open-prs

- handle-new-mails

Currently, these are handled in the "Update GitGitGadget's PRs" Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=2&_a=completed
and in the "Mirror Git List to GitGitGadget's PRs" Azure Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=5&_a=completed

The Pipeline definitions perform those pushes manually, and changing
CIHelper to do the push preemptively should allow for those definitions
to remain unchanged until the point in time when they are replaced by
workflows in gitgitgadget-workflows.

Eventually I'd like the same graceful notes ref pushing even for
script/update-mail-to-commit-notes.sh (which updates not the
`gitgitgadget`, but the `mail-to-commit` and `commit-to-mail` notes
refs), so that the "Update GitGitGadget's commit to mail notes" Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=9&_a=completed
can also be migrated to a GitHub workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>