misc-helper: Do push the notes ref as part of the common commands
I'd like the following `misc-helper` commands to call the new
`GitNotes.push()` method as needed, to benefit from the just-introduced
support for concurrent GitGitGadget workflow runs:

- update-open-prs

- update-commit-mappings

- handle-open-prs

- handle-new-mails

Currently, these are handled in the "Update GitGitGadget's PRs" Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=2&_a=completed
and in the "Mirror Git List to GitGitGadget's PRs" Azure Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=5&_a=completed

The Pipeline definitions perform those pushes manually, and changing
CIHelper to do the push preemptively should allow for those definitions
to remain unchanged until the point in time when they are replaced by
workflows in https://github.com/gitgitgadget/gitgitgadget-workflows.

Technically, this commit does not even touch `misc-helper.ts`; The
reason is that the logic is actually implemented in the `CIHelper`
class, which is a good thing, because that class will become the
principal entry-point of above-mentioned GitHub workflows.

Also note that there is a slight change of (`push`) behavior in
`handle-open-prs`: This command calls `CIHelper.handleOpenPRs()`, which
in turn calls `CIHelper.handlePR()` in a loop. Now, the notes ref will
be pushed every time the latter updates something, where before, the
notes ref was only pushed once, at the very end. This is arguably safer
because PRs that have already been handled in a Pipeline run that fails
somewhere in the middle won't be handled again upon the re-run.

Eventually I'd like the same graceful notes ref pushing even for
script/update-mail-to-commit-notes.sh (which updates not the
`gitgitgadget`, but the `mail-to-commit` and `commit-to-mail` notes
refs), so that the "Update GitGitGadget's commit to mail notes" Pipeline at
https://dev.azure.com/gitgitgadget/git/_build/index?definitionId=9&_a=completed
can also be migrated to a GitHub workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>