CIHelper: move the responsibility to `push` from `handleOpenPRs()` to `handlePR()`
We already push the notes ref when updating the PR metadata (i.e. the
object attached to the blob whose contents are the PR URL), therefore it
makes only sense to combine that with any update to the
GitGitGadgetOptions (i.e. the object attached to the empty object).

This simplifies the logic (and the separation of concerns) quite a bit.

Suggested-by: Chris. Webster <chris@webstech.net>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>