CIHelper: add a convenient function to initialize e.g. the Git worktree
In the upcoming GitHub Actions, we must assume that there is no
persistent Git worktree to use, unlike the situation of GitGitGadget's
Azure Pipelines, which use a single Git worktree that is retained and
reused across Pipeline runs.

To make this somewhat efficient, let's try to use a partial
(semi-shallow) clone.

There are a couple of other things we want to set up in the upcoming
GitHub Actions; Let's combine all of that setup in a single method of
the `CIHelper` class.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>