misc-helper: do configure the token for pushing the Git notes
When merging https://github.com/gitgitgadget/gitgitgadget/pull/1473 (Add
support for distributed notes updates), I did not anticipate that the
current Azure Pipelines were not prepared for that: Contrary to what I
had remembered, they did _not_ simply configure `http.extraHeader` to be
able to push (which is somewhat surprising because I authored them, and
I typically did that a lot in the olden days).

Happily, I _just_ introduced support for configuring a token in
`CIHelper` that is specifically used for accessing `gitgitgadget/git`,
including the Git notes pushes.

Let's teach `misc-helper` about this new trick, and pick up the token
that was configured via `set-app-token` and tell `CIHelper` to use it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>