patch-series: do accept GitGitGadget's push token explicitly
This completes the journey where the token is specified explicitly.

If you (like I did, when developing this patch) wonder how it worked
before: The Azure Pipelines set the `publishRemote` to a URL that
contains the token (which is also the reason why we need stunts like the
`redactGitHubToken()` method).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>