CIHelper.setupGitHubAction(): also process the SMTP inputs, if any
The upcoming `handle-pr-comment` Action wants to send emails, and for
that reason needs to accept SMTP credentials. These can be provided via
the `smtp-host`, `smtp-user` and `smtp-pass` Action inputs. Additional
options (that might be necessary to support projects other than Git) can
be passed in via the `smtp-opts` input, a JSON-formatted object to
augment/oeverride nodemailer's `SMTPTransport.Options` when sending the
emails.

Since most of GitGitGadget's Actions do not want to send emails, it is
silently ignored if these Action inputs are unset or empty when
`setupGitHubAction()` is called.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>