Add the `handle-pr-comment` GitHub Action
This is the culmination of all the preparation: The first GitHub Action
that will be used in a GitHub workflow that will replace these trusty
Azure Pipelines we used for so long:

  - GitGitGadget PR Handler:
    https://dev.azure.com/gitgitgadget/git/_build?definitionId=3

  - GitGitGadget PR Handler (git):
    https://dev.azure.com/gitgitgadget/git/_build?definitionId=13

  - GitGitGadget PR Handler (dscho):
    https://dev.azure.com/gitgitgadget/git/_build?definitionId=12

Note that this Action receives the full `pr-comment-url` as Action
input; Readers might wonder why the Action cannot receive the
information in a more fine-grained manner, seeing as the `github`
context of a GitHub workflow triggered by a new PR comment already has
all the necessary values. But this would require adding a
GitGitGadget-specific workflow to the main `git/git` repository,
something that was considered in
https://github.com/gitgitgadget/gitgitgadget/issues/861 but which would
be _extremely_ difficult to convince the core Git contributors to
accept. Therefore, the GitHub workflow using this Action will be a
`workflow_dispatch`-triggered one, for more details see
https://github.com/gitgitgadget-workflows/gitgitgadget-workflows/pull/1.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>