CIHelper: offer a convenient method to parse a PR comment URL
This will be used in the upcoming `handle-pr-comment` GitHub Action.

Note that the reason why this information needs to be passed to the
respective GitHub Action is that the corresponding GitHub workflows will
not run inside `gitgitgadget/git` and will therefore not automatically
receive the comment ID and PR number as part of the `github` context.
GitGitGadget's GitHub App will need to pass that information along when
receiving the webhook event that a PR comment has been created or
modified.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>