CIHelper: offer a convenient method to parse a PR URL
This will be used in the upcoming `handle-pr-push` GitHub Action.

Just like the `handle-pr-comment` Action, this Action will be used in a
GitHub workflow that runs in a different repository than the one where
the PR lives that is to be processed, therefore the information cannot
be passed in via the `github` context, and combined with ease of
debugging, I settled on the full PR URL as the unit of information that
the Action input should accept.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>