vscode: avoid high CPU usage via search
The search feature in VS Code is quite nice, except when it bogs down
the processor by trying to index files that aren't meant to be
searchable, such as `node_modules/`.

While at it, also exclude the same folders from the file watcher, and
restrict the often-rampant `ripgrep` processes to play nicely already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>