check-run: automatically update the conclusion in a `post` step
In a GitHub Action it is possible to register a `post` Action which will
run, if the main Action has been run in a GitHub workflow's job, as an
extra step after all the regular steps have been executed (or skipped).

This presents a fine opportunity to mark the the Check Run as completed
_automatically_, without the need to call the Action _again_.

By adding a new input called job status and filling it automatically
with -- you guessed it -- the job status, we can even automatically
determine the conclusion (I verified manually that this will receive the
expected value in the `post` Action).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>