Include the `LintCommit` configuration in `IConfig`
This way, the maximal number of columns can be configured freely per
project, via the project-specific config.

To reduce friction for on-boarding new projects by not having to look
in multiple locations when creating a config file, the definition of the
`ILintCommitConfig` interface is moved to `project-config.ts`, where all
the other project-specific interfaces live already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>