delete-ci-test-branches: add forgotten `async` keyword
When awaiting multiple asynchronous functions to complete via the
`Promise.all()` call, the assumption is that those functions are marked
up as `async`, too. Otherwise `typescript-eslint` v8.45.0 would complain
about `Promise.all()` being unnecessary, like this:

  lib/delete-ci-test-branches.ts
  Error:   117:9  error  Unexpected iterable of non-Promise
    (non-"Thenable") values passed to promise aggregator
    @typescript-eslint/await-thenable

The underlying reason why that function was not marked up as `async` is
that the `graphql()` call returns `GraphQlResponse<ResponseData>`, which
is not immediately identified as a `Promise` (but it is).

So let's `await` that call, _and_ mark the function as `async`, _before_
upgrading to `typescript-eslint` v8.45.0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>